+++ packages/core/src/index.ts
@@ -9,6 +9,7 @@ export * from "./domain/errors.js";
 export * from "./domain/plugins.js";
 export * from "./domain/project.js";
 export * from "./domain/sources.js";
+export * from "./domain/supervision.js";
 export * from "./domain/version.js";
 export * from "./ports/backup.js";
 export * from "./ports/plugin-catalog.js";